add haversine heuristic. scale heuristic factor to match real scale (multiply by 10)

+++ parte-2/include/graph_utils.hpp
@@ -2,8 +2,8 @@
 #include <vector>
 
 struct Coord {
-  double lon;
-  double lat;
+  int lon;
+  int lat;
 };
 
 class Graph {
